renamed dir and updated imports
