+++ test/users/2_readUser.js
@@ -1,4 +1,4 @@
-import { User } from "../../schemas/user.js";
+import { User } from "../../models/user.js";
 import assert from "assert";
 
 describe(`Reading a user's details from MongoDB`, () => {
